[DIDA] feat. build-logic > Data모듈 관련 의존성 추가
