+++ build-logic/convention/src/main/java/com/example/convention/AndroidDataConventionPlugin.kt
@@ -22,6 +22,30 @@ class AndroidDataConventionPlugin : Plugin<Project> {
             val libs = extensions.getByType<VersionCatalogsExtension>().named("libs")
             dependencies{
                 add("implementation", project(":domain"))
+
+                add("implementation", libs.findLibrary("androidx-core").get())
+                add("implementation", libs.findLibrary("junit-junit").get())
+                add("implementation", libs.findLibrary("androidx-test-junit").get())
+
+                // Network
+                add("implementation", libs.findBundle("gson").get())
+                add("implementation", libs.findLibrary("squareup-retrofit2").get())
+                add("implementation", libs.findLibrary("squareup-okhttp").get())
+                add("implementation", libs.findLibrary("squareup-okhttp-interceptor").get())
+
+                // Coroutine
+                add("implementation", libs.findLibrary("kotlin-coroutine-core").get())
+
+                // Room
+                add("kapt", libs.findLibrary("androidx-room-compiler").get())
+                add("implementation", libs.findBundle("room").get())
+
+                // Moshi
+                add("implementation", libs.findBundle("moshi").get())
+
+                // DataStore
+                add("implementation", libs.findBundle("datastore").get())
+
             }
 
         }
