--- build-logic/convention/src/main/java/com/example/convention/AndroidDomainConventionPlugin.kt
@@ -1,26 +1,0 @@
-package com.example.convention
-
-import com.android.build.gradle.LibraryExtension
-import com.example.convention.project.configureKotlinAndroid
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-import org.gradle.api.artifacts.VersionCatalogsExtension
-import org.gradle.kotlin.dsl.configure
-import org.gradle.kotlin.dsl.dependencies
-import org.gradle.kotlin.dsl.getByType
-
-class AndroidDomainConventionPlugin : Plugin<Project> {
-    override fun apply(target: Project) {
-        with(target){
-            with(pluginManager){
-                apply("android.library")
-                apply("android.hilt")
-            }
-
-            val libs = extensions.getByType<VersionCatalogsExtension>().named("libs")
-            dependencies{
-            }
-
-        }
-    }
-}
