+++ domain/build.gradle.kts
@@ -1,6 +1,4 @@
 plugins {
-//    id("android.domain")
-
     id("java-library")
     id("org.jetbrains.kotlin.jvm")
     id("kotlin-kapt")
